Request parsing capabilities moved to HttpBackend
